requirements.txt
torch
torchvision
albumentations
opencv-python
Pillow
Flask
numpy
